feat: Add debounce utility function
This commit adds a new utility function called `debounce` to the `src/lib/utils/debounce.ts` file. The `debounce` function is used to delay the execution of a callback function by a specified amount of time. It improves the codebase by providing a reusable and efficient solution for handling debounce functionality.