feat: Add POST endpoint for inserting general observations
This commit adds a new POST endpoint to the server file in the "generalObservations" endpoints directory. The endpoint allows authenticated users to insert general observations into the "General_Observations" table in the database. It validates the user's session, checks for authorization, and inserts the data into the table using the Supabase library. If there is an error, it throws an appropriate error response. This enhancement improves the functionality of the application by providing a way to store and retrieve general observations from the database.