refactor: Update +page.server.ts to load general observations data
This commit updates the +page.server.ts file to load general observations data from the 'General_Observations' table in the database. It adds a new dependency on 'supabase:db:General_Observations' and fetches the data using the Supabase library. If there is an error, it throws an appropriate error response. This refactor enhances the functionality of the page by including general observations data in the response.